Extract useful function to a common helper, and move the rest to legacy.
