+++ test/datatype.unittest.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 #
 # License: MIT
-# Last Change: Thu Sep 20, 2018 at 05:23 PM -0400
+# Last Change: Thu Dec 06, 2018 at 04:48 PM -0500
 
 import unittest
 
@@ -9,7 +9,6 @@ import sys
 sys.path.insert(0, '..')
 
 from pyUTM.datatype import ColNum, range
-from pyUTM.datatype import BrkStr
 from pyUTM.datatype import GenericNetNode
 from pyUTM.datatype import ExcelCell
 
@@ -66,52 +65,6 @@ class DataTypeTester(unittest.TestCase):
     def test_range_generation(self):
         self.assertEqual(range(ColNum('A'), ColNum('E')), [1, 2, 3, 4])
         self.assertEqual(str(range(ColNum('A'), ColNum('E'))[1]), 'B')
-
-
-class BrkStrTester(unittest.TestCase):
-    def test_str_basic_function(self):
-        name = BrkStr('name')
-        self.assertEqual(name, 'name')
-        self.assertEqual(name + '_something', 'name_something')
-
-    def test_iteration(self):
-        name = BrkStr('name')
-        self.assertEqual(list(name), ['n', 'a', 'm', 'e'])
-
-    def test_lack_of_representation(self):
-        name = BrkStr('name')
-        name_processed = list(name)
-        self.assertEqual(isinstance(name_processed[0], str), True)
-        self.assertEqual(isinstance(name_processed[0], BrkStr), False)
-
-    def test_split_signal_id(self):
-        self.assertEqual(
-            BrkStr.split_signal_id_into_three('JP0_JT11_SOMETHING'),
-            ['JP0', 'JT11', 'SOMETHING']
-        )
-        self.assertEqual(
-            BrkStr.split_signal_id_into_three('JP0_JT11_SOMETHING_ELSE'),
-            ['JP0', 'JT11', 'SOMETHING_ELSE']
-        )
-        self.assertEqual(
-            BrkStr.split_signal_id_into_three('JP0_JT11_SOMETHING_ELSE_IF'),
-            ['JP0', 'JT11', 'SOMETHING_ELSE_IF']
-        )
-
-    def test_overloaded_contain(self):
-        name1 = BrkStr('JP1_JD12_SOME_SOME_ELSE')
-        name2 = BrkStr('JP11_JP12_SOME_SOME_ELSE')
-        name3 = BrkStr('JD11_JPL2_1V5_M')
-
-        self.assertTrue('JP1' in name1)
-        self.assertFalse('JP1' in name2)
-        self.assertTrue('JP11' in name2)
-        self.assertTrue('JP12' in name2)
-        self.assertTrue('SOME_SOME_ELSE' in name1)
-        self.assertTrue('SOME_SOME_ELSE' in name2)
-        self.assertTrue('JD11' in name3)
-        self.assertTrue('JPL2' in name3)
-        self.assertFalse('JPL3' in name3)
 
 
 class ExcelCellTester(unittest.TestCase):
